backend/pyproject.toml
[project]
name = "ghs-carnival-backend"
version = "0.1.0"
description = "FastAPI backend for GHS Carnival"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6",
  "pydantic>=2.9.2",
  "python-dotenv>=1.0.1",
  "prisma>=0.15.0",
]

[tool.uvicorn]
factory = false
